Conflicts:
	src/emc/motion/command.c
	src/emc/task/emccanon.cc
	tcl/bin/emccalib.tcl
	tests/halui/jogging/halui.ini

+++ src/emc/nml_intf/emc.hh
@@ -568,7 +568,7 @@ int emcSetMaxFeedOverride(double maxFeedScale);
 int emcSetupArcBlends(int arcBlendEnable,
         int arcBlendFallbackEnable,
         int arcBlendOptDepth,
-        double arcBlendGapCycles,
+        int arcBlendGapCycles,
         double arcBlendRampFreq);
 
 extern int emcUpdate(EMC_STAT * stat);
